refactor(atm-detail): move atm-detail top of app and move params from OrderContact to ContactContext
