+++ app/atm-detail/layout.tsx
@@ -3,6 +3,7 @@
 import { OrderDetailProvider } from '@/context/OrderDetailContext'
 import { useSearchParams } from 'next/navigation'
 import { PropsWithChildren } from 'react'
+import { ContactProvider } from '@/context/ContactContext'
 
 type RootLayoutProps = PropsWithChildren
 
@@ -10,8 +11,10 @@ export default function RootLayout({ children }: Readonly<RootLayoutProps>) {
   const searchParams = useSearchParams()
   const id = searchParams.get('id') || ''
   return (
-    <OrderDetailProvider id={id}>
-      <main className="flex h-full min-h-screen flex-col">{children}</main>
-    </OrderDetailProvider>
+    <ContactProvider>
+      <OrderDetailProvider id={id}>
+        <main className="flex h-full min-h-screen flex-col">{children}</main>
+      </OrderDetailProvider>
+    </ContactProvider>
   )
 }
