+++ services/order.ts
@@ -167,7 +167,7 @@ export const addOrder = async (order: OrderData): Promise<void> => {
     discount_title: order.discount_title || '',
     discount_code: order.discount_code || '',
     gobackurl: order.paystatus?.includes('atm')
-      ? `https://${location.host}/member/orders/atm-detail/checkout`
+      ? `https://${location.host}/atm-detail/checkout`
       : `https://${location.host}/confirm-order/result`,
     CVSAddress: order.CVSAddress || '',
     CVSStoreID: order.CVSStoreID || '',
